start implementing ens_cloud_arrival_time
write a set of tests for cloud_arrival_time works

+++ tests/medium/pyflexplot/io/test_read_field_ensemble.py
@@ -271,10 +271,7 @@ class TestReadFieldEnsemble_Multiple:
             "max": np.nanmax,
             "thr_agrmt": (
                 lambda arr, axis: threshold_agreement(
-                    arr,
-                    self.agreement_threshold_concentration,
-                    axis=axis,
-                    dtype=arr.dtype,
+                    arr, self.agreement_threshold_concentration, axis=axis,
                 )
             ),
         }[ens_var]
